State copy should clone dictionaries
When copying a state, the visitCounts and turnIndices objects need to be cloned, so that later changes don't change their entries. Otherwise, you get strange effects, such as with the following ink source:

-> bug
=== bug
    Removing this line resolves the issue.
    {not knot:  
         // Including a line here resolves the issue
        -> knot ->
    - else: 
        This content should NEVER be seen.
    }
    -> END
    
=== knot
    This content SHOULD be seen.
    -> END

Note there may be other members of the State object which need cloning as well, but my comparison with the c# suggested this is all that's needed?