Update PRNG logic
The new algorithm, Mulberry32, doesn't return its internal state as
the generated random number. This means that previousRandom, the value
we keep between generations, should not be the actual previous random
number generated, but the previous seed normally kept inside the
generator. To do that, a new nextSeed() method is added.

Also, the storySeed is no longer fed into the PRNG, but only used on
first initialization.