Replace legacy Jasmine dependency in tests with Jest builtin
The global `jasmine` API has been removed from Jest with the new test
runner `jest-circus` introduced in v27–v29. While possible to import the
legacy Jasmine package, it seemed to be more robust to upgrade to the
`jest.fn` API which behaves exactly the same as `jasmine.createSpy`.

The main difference is the `callThrough()` method in Jasmine spies. This
behaviour is the default for `jest.fn()`, so the upgrade works as-is.